# cmd/api/docs/swagger.yaml
definitions:
...
  models.CreateOffer:
    properties:
      categoryID:
        type: integer
      discount:
        type: integer
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /admin/offers:
    get:
      consumes:
      - application/json
      description: Admin can view the list of  offers
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List Offers
      tags:
      - Admin
  /admin/offers/create:
    post:
      consumes:
      - application/json
      description: Admin can add new  offers
      parameters:
      - description: Offer
        in: body
        name: offer
        required: true
        schema:
          $ref: '#/definitions/models.CreateOffer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Add Offer
      tags:
      - Admin
  /admin/offers/expire:
    post:
      consumes:
      - application/json
      description: Admin can add Expire  offers
      parameters:
      - description: Category ID
        in: query
        name: catID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Expire Offer
      tags:
      - Admin
...
swagger: "2.0"